+++ app/protected/modules/contacts/controllers/ContactApiController.php
@@ -62,6 +62,11 @@
             try
             {
                 $model = Contact::getById($id);
+                $isAllowed = ControllerSecurityUtil::resolveAccessCanCurrentUserReadModel($model);
+                if ($isAllowed === false)
+                {
+                    throw new Exception('This action is not allowed.');
+                }
                 $util  = new RedBeanModelToApiDataUtil($model);
                 $data  = $util->getData();
                 $outputArray = array();
@@ -205,12 +210,14 @@
                 }
                 else
                 {
+                    $outputArray['data'] = null;
                     $outputArray['status'] = 'FAILURE';
                     $outputArray['message'] = Yii::t('Default', 'Model could not be saved.');
                 }
             }
             catch (Exception $e)
             {
+                $outputArray['data'] = null;
                 $outputArray['status'] = 'FAILURE';
                 $outputArray['message'] = $e->getMessage();
             }
@@ -222,7 +229,11 @@
             try
             {
                 $model = Contact::getById($id);
-
+                $isAllowed = ControllerSecurityUtil::resolveAccessCanCurrentUserWriteModel($model);
+                if ($isAllowed === false)
+                {
+                    throw new Exception('This action is not allowed.');
+                }
                 if (isset($data['firstName']))
                 {
                     $model->firstName            = $data['firstName'];
@@ -430,12 +441,14 @@
                 }
                 else
                 {
+                    $outputArray['data'] = null;
                     $outputArray['status'] = 'FAILURE';
                     $outputArray['message'] = Yii::t('Default', 'Model could not be saved.');
                 }
             }
             catch (Exception $e)
             {
+                $outputArray['data'] = null;
                 $outputArray['status'] = 'FAILURE';
                 $outputArray['message'] = $e->getMessage();
             }
@@ -447,6 +460,11 @@
             try
             {
                 $model = Contact::getById($id);
+                $isAllowed = ControllerSecurityUtil::resolveAccessCanCurrentUserDeleteModel($model);
+                if ($isAllowed === false)
+                {
+                    throw new Exception('This action is not allowed.');
+                }
                 $model->delete();
                 $outputArray['status'] = 'SUCCESS';
                 $outputArray['message'] = '';
