+++ app/protected/modules/opportunities/controllers/OpportunityApiController.php
@@ -45,12 +45,14 @@
                 }
                 else
                 {
+                    $outputArray['data'] = null;
                     $outputArray['status'] = 'FAILURE';
                     $outputArray['message'] = Yii::t('Default', 'Error');
                 }
             }
             catch (Exception $e)
             {
+                $outputArray['data'] = null;
                 $outputArray['status'] = 'FAILURE';
                 $outputArray['message'] = $e->getMessage();
             }
@@ -62,6 +64,11 @@
             try
             {
                 $model = Opportunity::getById($id);
+                $isAllowed = ControllerSecurityUtil::resolveAccessCanCurrentUserReadModel($model);
+                if ($isAllowed === false)
+                {
+                    throw new Exception('This action is not allowed.');
+                }
                 $util  = new RedBeanModelToApiDataUtil($model);
                 $data  = $util->getData();
                 $outputArray = array();
@@ -140,12 +147,14 @@
                 }
                 else
                 {
+                    $outputArray['data'] = null;
                     $outputArray['status'] = 'FAILURE';
                     $outputArray['message'] = Yii::t('Default', 'Model could not be saved.');
                 }
             }
             catch (Exception $e)
             {
+                $outputArray['data'] = null;
                 $outputArray['status'] = 'FAILURE';
                 $outputArray['message'] = $e->getMessage();
             }
@@ -157,7 +166,11 @@
             try
             {
                 $model = Opportunity::getById($id);
-
+                $isAllowed = ControllerSecurityUtil::resolveAccessCanCurrentUserWriteModel($model);
+                if ($isAllowed === false)
+                {
+                    throw new Exception('This action is not allowed.');
+                }
                 if (isset($data['name']))
                 {
                     $model->name            = $data['name'];
@@ -240,12 +253,14 @@
                 }
                 else
                 {
+                    $outputArray['data'] = null;
                     $outputArray['status'] = 'FAILURE';
                     $outputArray['message'] = Yii::t('Default', 'Model could not be saved.');
                 }
             }
             catch (Exception $e)
             {
+                $outputArray['data'] = null;
                 $outputArray['status'] = 'FAILURE';
                 $outputArray['message'] = $e->getMessage();
             }
@@ -257,6 +272,11 @@
             try
             {
                 $model = Opportunity::getById($id);
+                $isAllowed = ControllerSecurityUtil::resolveAccessCanCurrentUserDeleteModel($model);
+                if ($isAllowed === false)
+                {
+                    throw new Exception('This action is not allowed.');
+                }
                 $model->delete();
                 $outputArray['status'] = 'SUCCESS';
                 $outputArray['message'] = '';
