+++ app/protected/modules/tasks/controllers/TaskApiController.php
@@ -45,12 +45,14 @@
                 }
                 else
                 {
+                    $outputArray['data'] = null;
                     $outputArray['status'] = 'FAILURE';
                     $outputArray['message'] = Yii::t('Default', 'Error');
                 }
             }
             catch (Exception $e)
             {
+                $outputArray['data'] = null;
                 $outputArray['status'] = 'FAILURE';
                 $outputArray['message'] = $e->getMessage();
             }
@@ -62,6 +64,11 @@
             try
             {
                 $model = Task::getById($id);
+                $isAllowed = ControllerSecurityUtil::resolveAccessCanCurrentUserReadModel($model);
+                if ($isAllowed === false)
+                {
+                    throw new Exception('This action is not allowed.');
+                }
                 $util  = new RedBeanModelToApiDataUtil($model);
                 $data  = $util->getData();
                 $outputArray = array();
@@ -126,12 +133,14 @@
                 }
                 else
                 {
+                    $outputArray['data'] = null;
                     $outputArray['status'] = 'FAILURE';
                     $outputArray['message'] = Yii::t('Default', 'Model could not be saved.');
                 }
             }
             catch (Exception $e)
             {
+                $outputArray['data'] = null;
                 $outputArray['status'] = 'FAILURE';
                 $outputArray['message'] = $e->getMessage();
             }
@@ -143,7 +152,11 @@
             try
             {
                 $model = Task::getById($id);
-
+                $isAllowed = ControllerSecurityUtil::resolveAccessCanCurrentUserWriteModel($model);
+                if ($isAllowed === false)
+                {
+                    throw new Exception('This action is not allowed.');
+                }
                 if (isset($data['name']))
                 {
                     $model->name     = $data['name'];
@@ -203,12 +216,14 @@
                 }
                 else
                 {
+                    $outputArray['data'] = null;
                     $outputArray['status'] = 'FAILURE';
                     $outputArray['message'] = Yii::t('Default', 'Model could not be saved.');
                 }
             }
             catch (Exception $e)
             {
+                $outputArray['data'] = null;
                 $outputArray['status'] = 'FAILURE';
                 $outputArray['message'] = $e->getMessage();
             }
@@ -220,6 +235,11 @@
             try
             {
                 $model = Task::getById($id);
+                $isAllowed = ControllerSecurityUtil::resolveAccessCanCurrentUserDeleteModel($model);
+                if ($isAllowed === false)
+                {
+                    throw new Exception('This action is not allowed.');
+                }
                 $model->delete();
                 $outputArray['status'] = 'SUCCESS';
                 $outputArray['message'] = '';
